修复数据库模型类型匹配问题，确保与现有数据库结构兼容
- 修复所有主键类型从int64改为int以匹配MySQL的INT类型
- 修复外键约束不匹配问题
- 统一使用Moment模型替代Post模型以保持兼容性
- 完善搜索服务实现，使用真实的SearchHistory表
- 保持向后兼容性，不改变现有数据库结构

+++ internal/models/verification.go
@@ -18,7 +18,7 @@ type VerificationCode struct {
 
 // SearchHistory 搜索历史模型
 type SearchHistory struct {
-	ID        int64     `json:"id" gorm:"primaryKey;autoIncrement"`
+	ID        int       `json:"id" gorm:"primaryKey;autoIncrement"`
 	UserID    string    `json:"userId" gorm:"type:char(10);not null;index"`
 	Keyword   string    `json:"keyword" gorm:"type:varchar(100);not null"`
 	CreatedAt time.Time `json:"createdAt"`
